Separate ApiBootstrap into it's own package
- Can now be used by all the different apps (code-studio, embed-grid, embed-chart)
- Could update the error shown to use some of our fancier components (uses modal)
- Separate JS API types into their own package - now you can use the types without importing jsapi-shim
