+++ packages/code-studio/tsconfig.json
@@ -6,17 +6,8 @@
     "noEmit": true,
     "emitDeclarationOnly": false
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    "src/**/*.js",
-    "src/**/*.jsx"
-  ],
-  "exclude": [
-    "node_modules",
-    "src/**/*.test.*",
-    "src/**/__mocks__/*"
-  ],
+  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.js", "src/**/*.jsx"],
+  "exclude": ["node_modules", "src/**/*.test.*", "src/**/__mocks__/*"],
   "references": [
     {
       "path": "../chart"
@@ -46,7 +37,16 @@
       "path": "../iris-grid"
     },
     {
+      "path": "../jsapi-bootstrap"
+    },
+    {
       "path": "../jsapi-shim"
+    },
+    {
+      "path": "../jsapi-types"
+    },
+    {
+      "path": "../jsapi-utils"
     },
     {
       "path": "../log"
