+++ packages/jsapi-bootstrap/src/useApi.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ApiContext } from './ApiBootstrap';
+
+export const useApi = () => useContext(ApiContext);
+
+export default useApi;
