+++ packages/jsapi-types/README.md
@@ -1,0 +1,17 @@
+# @deephaven/jsapi-types
+
+Manually created type definitions of the Deephaven JS API.
+
+## Install
+
+```bash
+npm install --save @deephaven/jsapi-types
+```
+
+## Usage
+
+```javascript
+import { Column } from '@deephaven/jsapi-types'
+
+...
+```
