+++ packages/jsapi-types/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@deephaven/jsapi-shim",
+  "name": "@deephaven/jsapi-types",
   "version": "0.29.0",
-  "description": "Deephaven JSAPI Shim",
+  "description": "Deephaven JSAPI Types",
   "author": "Deephaven Data Labs LLC",
   "license": "Apache-2.0",
   "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/deephaven/web-client-ui.git",
-    "directory": "packages/jsapi-shim"
+    "directory": "packages/jsapi-types"
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -19,10 +19,6 @@
   "scripts": {
     "build": "cross-env NODE_ENV=production run-p build:*",
     "build:babel": "babel ./src --out-dir ./dist --extensions \".ts,.tsx,.js,.jsx\" --source-maps --root-mode upward"
-  },
-  "dependencies": {
-    "@deephaven/jsapi-types": "file:../jsapi-types",
-    "prop-types": "^15.7.2"
   },
   "devDependencies": {
     "@deephaven/tsconfig": "file:../tsconfig"
